Fetching Players Info
Have a working version of fetching all the nba player info. Currentlythe
code displays the fetched data in a pandas data frame. Next stepswould
be to add the code to insert all the fetched data into the databseand
update the routes to pull from the database instead of the nba_api.Also
added a progress bar to show progress when fetching data andupdated the
player model with the correct columns.