Player Game Logs
Created a script to fetch and insert all the player game logs for a nba
season into the database. Added an API endpoint to retrieve the players
last 5 games. Created a function to query the database to retrieve a
number of games in a specific season and created a function that will
automatically generate the season_year to get the latest season. Started
implementing type hinting to make debugging, reviewing, and documenting
code easier. Able to display the last 5 recent games from the frontend
of the application.