Players Page
Updated the code in the players page to display the correct types for
typescript and display an error if there is an issue with fetching
players. Also updated the columns to display the players number,
position, school, country and players picture next to their full name.
The layout or table may change.